Finding the quickest way to calculate.
Here's a sample from the workbook I am currently working on. I have doubts with the formula I have used in the cell H15. Everytime I write a number in the D15 cell, it shows nothing in the cell H15. 
The above table has the formulas i used in the workbook and second table has the formulas I wrote to make calculations faster but as it turned out it didnt work.
Please check the attached file and if best way possible to solve the problem.